requirements.txt
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
PyYAML>=6.0
lxml>=4.9.0
yfinance>=0.2.0
alpha_vantage>=2.3.1 tqdm
pyarrow
